test/tempfile.h
Moved TempFile from test/cluster.cpp to a header file of its own so that
it can be utilized in other unit-tests too.
